Port RegistrationBuilderCompatability tests to System.Composition.TypeParts where it belongs
