+++ src/System.Composition/tests/System.Composition.Tests.csproj
@@ -36,7 +36,6 @@
     <Compile Include="OpenGenericsTests.cs" />
     <Compile Include="OptionalImportTests.cs" />
     <Compile Include="PropertyExportTests.cs" />
-    <Compile Include="RegistrationBuilderCompatibilityTests.cs" />
     <Compile Include="SharingTests.cs" />
     <Compile Include="Util\AssertX.cs" />
   </ItemGroup>
